+++ policy/vmss-package.tf
@@ -1,0 +1,7 @@
+resource "checkpoint_management_package" "vmss" {
+  name              = "VMSS"
+  comments          = "Policy for VMSS gateways"
+  color             = "blue"
+  threat_prevention = false
+  access            = true
+}